+++ torrentfile/commands.py
@@ -1,16 +1,21 @@
 #! /usr/bin/python3
 # -*- coding: utf-8 -*-
 
-#####################################################################
-# THE SOFTWARE IS PROVIDED AS IS WITHOUT WARRANTY OF ANY KIND,
-# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
-# OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-# NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
-# HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
-# WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
-# OTHER DEALINGS IN THE SOFTWARE.
-#####################################################################
+##############################################################################
+#    Copyright (C) 2021-current alexpdev
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+##############################################################################
 """
 The commands module contains the Action Commands executed by the CLI script.
 
@@ -65,9 +70,10 @@ def create(args: list):
         "private": args.private,
         "outfile": args.outfile,
         "comment": args.comment,
+        "cwd": args.cwd,
     }
-
-    logger.debug("Program has entered torrent creation mode.")
+    print(args, kwargs)
+    logger.debug("Create new torrent file from %s", args.content)
 
     if args.meta_version == "2":
         torrent = TorrentFileV2(**kwargs)
@@ -75,8 +81,8 @@ def create(args: list):
         torrent = TorrentFileHybrid(**kwargs)
     else:
         torrent = TorrentFile(**kwargs)
-    logger.debug("Completed torrent files meta info assembly.")
     outfile, meta = torrent.write()
+    logger.debug("Torrent file creation complete.")
 
     if args.magnet:
         magnet(outfile)