+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # TorrentFile
 
+## Version 0.7.1
+
+- split CI integration into separate platform specific files
+- added new cli argument `--cwd` which changes the default save
+to location to the current working directory (this will be default in future)
+- added unit tests to cover the new argument
+- Changed license to a the more permissive Apache 2 software license
+
+---------------------
+
 ## Version 0.7.0
 
 - Fixed issues with logging to file.